add `git pull` functionality
- simplify config.ini and remove unneeded items
- add shell script that executes `git pull`
- improved HTTP response

Note:
`git pull` is not executed within php via [shell_]exec(...) due to
issues with the ssh-agent (that is not permanently running on the
webhosting server).
Preconditions for pulling are a running ssh-agent with the right key
already added.
By doing this step by step focus is lost and the ssh-agent was killed.
Therefore the script does the work at once.